Standardize emote delete action
